requirements.txt
# Core
langgraph>=0.2.0
langchain-core>=0.3.0
mcp>=1.0.0

# MLX
mlx>=0.22.0
mlx-lm>=0.21.0

# Utils
pydantic>=2.6.0
python-dotenv>=1.0.1
